requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.20
APScheduler==3.10.4
python-dotenv==1.0.1
qrcode==7.4.2
Pillow==10.4.0
email-validator==2.2.0
gunicorn==21.2.0
